move conditional AES256GCM availability to runtime
whether or not this algorithm is supported by the cpu is a runtime property.
there is no need for #ifdefs around datastructures, nor conditional
compilation.

add a static variable to be set by libsodium-provided runtime check at module
boot time, indicating whether the algorithm is supported on the current system.
that will keep the performance impact minimal in the functions that need to
check.